requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
sqlalchemy
